rename: routing 표현 bottom tab navigation으로 변경
