+++ ReactNativeStudy/Zero/Week1/Todo/App.tsx
@@ -7,7 +7,7 @@
 
 import React from 'react';
 import { RecoilRoot } from 'recoil';
-import Routing from './src/utility/Routing';
+import Routing from './src/utility/BottomTabNavigation';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import StackNavigation from './src/navigation/StackNavigation';
